+++ js/app.js
@@ -31,7 +31,6 @@ import WindowView from "../modules/window/view";
 import SidebarView from "../modules/sidebar/view";
 import ShadowView from "../modules/tools/shadow/view";
 import ParcelSearchView from "../modules/tools/parcelSearch/view";
-import SearchByCoordView from "../modules/tools/searchByCoord/view";
 import LineView from "../modules/tools/pendler/lines/view";
 import AnimationView from "../modules/tools/pendler/animation/view";
 import FilterView from "../modules/tools/filter/view";
@@ -236,8 +235,8 @@ async function loadApp () {
                 new ParcelSearchView({model: tool});
                 break;
             }
-            case "searchByCoord": {
-                new SearchByCoordView({model: tool});
+            case "saveSelection": {
+                new SaveSelectionView({model: tool});
                 break;
             }
             /**