+++ src/modules/tools/stateTools.js
@@ -1,6 +1,7 @@
 import Draw from "./draw/components/Draw.vue";
 import FileImport from "./fileImport/components/FileImport.vue";
 import Gfi from "./gfi/components/Gfi.vue";
+import SearchByCoord from "./searchByCoord/components/SearchByCoord.vue";
 import Measure from "./measure/components/Measure.vue";
 import SaveSelection from "./saveSelection/components/SaveSelection.vue";
 import ScaleSwitcher from "./scaleSwitcher/components/ScaleSwitcher.vue";
@@ -17,6 +18,7 @@ const state = {
         draw: Draw,
         fileImport: FileImport,
         gfi: Gfi,
+        searchByCoord: SearchByCoord,
         measure: Measure,
         saveSelection: SaveSelection,
         scaleSwitcher: ScaleSwitcher,